인증게시판수정,삭제1

+++ src/main/java/com/iso/plogues/proof/file/model/service/ProofFileService.java
@@ -8,6 +8,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.iso.plogues.exception.FailedDeleteException;
 import com.iso.plogues.exception.FailedInsertException;
+import com.iso.plogues.exception.FileUploadException;
 import com.iso.plogues.proof.file.model.dao.ProofFileMapper;
 import com.iso.plogues.util.file.File;
 import com.iso.plogues.util.file.FileDto;
@@ -57,15 +58,15 @@ public class ProofFileService {
 	}
 
 	@Transactional
-	public void updateFile(List<MultipartFile> files, Long proofNo) {
+	public void updateFile(List<MultipartFile> files, Long proofNo, String boardType) {
+		if (files == null || files.size() != 2) {
+			throw new FileUploadException("인증 사진은 2장을 등록해야 합니다.");
+		}
 
 		deleteFile(proofNo);
 
 		saveProofFiles(files, proofNo);
 
 	}
-	
-	
-	
-	
+
 }
